Restore nick upon reconnect
Make whitenames happy, as their nicks do not actually get stored in any permanent location server-side. So if the server cycles then it will either generate a new new or use the one that's still in the ws connection string, which may not be the one that's still in use.